chore: move buttons to NavSidebar
